move memory test
